Delete IBPreview namespace
