--- Sources/UIViewKit/IBPreviews/IBPreview.swift
@@ -1,13 +1,0 @@
-//
-//  IBPreview.swift
-//  UIViewKit
-//
-//  Created by Blazej SLEBODA on 10/02/2024.
-//
-
-import Foundation
-
-public final class IBPreview {
-
-    private init() { }
-}
